Moved ImageID enum to UiImageManager
