+++ taskey/src/taskey/ui/UiMain.java
@@ -13,10 +13,10 @@ import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
 import taskey.constants.UiConstants;
-import taskey.constants.UiConstants.ImageID;
 import taskey.logger.TaskeyLog;
 import taskey.logger.TaskeyLog.LogSystems;
 import taskey.ui.utility.UiImageManager;
+import taskey.ui.utility.UiImageManager.ImageID;
 
 /**
  * @@author A0125419H
